+++ MultiplicationTree.h
@@ -21,7 +21,7 @@ class MultiplicationTree
         int delay(shared_ptr<Node>);
         string expression(shared_ptr<Node>);
         vector<OperationNode*> cost(shared_ptr<Node>);
-        __int128 execute(shared_ptr<Node>, __int128, __int128);
+        long long execute(shared_ptr<Node>, long long, long long);
 
     public:
         MultiplicationTree();
@@ -33,7 +33,7 @@ class MultiplicationTree
         string getCost();
         void setRoot(shared_ptr<OperationNode>);
 	void setDescription(string);	
-        __int128 executeMultiplication(__int128, __int128);
+        long long executeMultiplication(long long, long long);
         virtual ~MultiplicationTree() = default;
 };
